refactor: restructure FilmModal components for better separation of concerns
- Rename WatchedFields/WatchlistFields to WatchedModalContent/WatchlistModalContent
- Move data fetching, form state management, and update logic from FilmModal to individual modal content components
- FilmModal now only handles modal display control and delegates content to specialized components
- Each modal content component manages its own data lifecycle and includes FilmModalHeader
- Remove console.log from useFilmData hook
- Clean up imports and simplify component interfaces

This improves maintainability by giving each component a single, clear responsibility.
