+++ src/hooks/useFilmData.ts
@@ -5,7 +5,6 @@ import { TMDBFilmData } from "@/types";
 
 export const useFilmData = (filmId?: number) => {
     const [filmData, setFilmData] = useState<TMDBFilmData | null>(null);
-    console.log("useFilmData called with filmId:", JSON.stringify(filmData));
 
     const fetchFilmData = useCallback(async (id: number) => {
         const response = await fetch(`${TMDB_API_URL}/movie/${id}?language=en-US&api_key=${TMDB_API_KEY}`, { method: "GET" });
